Time: 1313 ms (5.16%) | Memory: 57 MB (91.67%) - LeetSync

+++ 2483-task-scheduler-ii/task-scheduler-ii.java
@@ -2,16 +2,16 @@ class Solution {
     public long taskSchedulerII(int[] tasks, int space) 
     {
         long time=0;
-        HashMap<Long,Long> map=new HashMap<>();
+        HashMap<Integer,Long> map=new HashMap<>();
         for(int task : tasks)
         {
-            if(map.containsKey((long)task))
-            time=Math.max(map.get((long)task)+space,time);
+            if(map.containsKey(task))
+            time=Math.max(map.get(task)+space,time);
             // Complete the current task
             time++;
             // Map the task to its completion time.
             System.out.println(task+"-"+time);
-            map.put((long)task,time);
+            map.put(task,time);
         }
         return time;
     }
